+++ src/qpe_toolbox/circuit/parametrized_circuits.py
@@ -292,18 +292,18 @@ def generate_brickwall_circuit(
 
     Circuit structure (one layer, ``start_ent=False``)::
 
-        q0 ──[]────●─────────
+        q0 ──[]───●───────
                   │
-        q1 ──[]────●───●─────
+        q1 ──[]───●───●───
                       │
-        q2 ──[]────●───●─────
-                  │
-        q3 ──[]────●───●─────
-                      │
-        q4 ──[]────●───●─────
-                  │
-        q5 ──[]────●─────────
-          <──────────────>
+        q2 ──[]───●───●───
+                  |
+        q3 ──[]───●───●───
+                      |
+        q4 ──[]───●───●───
+                  |
+        q5 ──[]───●───────
+          <─────────────>
             first layer
 
     where::
